CMakeLists.txt: Change target_compile_definition
If target_compile_definition is using PUBLIC or INTERFACE the used
compile flags are added to the cmake target installed on the host.

If multiple static libraries for the different dilithium forms are
linked the warnings shown below are emitted by the compiler.

[22/258] Building CXX object tests/unit/CMakeFiles/......
<command-line>: warning: "DILITHIUM_MODE" redefined
<command-line>: note: this is the location of the previous definition
<command-line>: warning: "DILITHIUM_MODE" redefined
<command-line>: note: this is the location of the previous definition

These warning can be omitted when the compile definitions are set to
PRIVATE.