CMakeLists.txt: Add cmake install target
Previously it was not possible to use cmake to generate rules for
installing the library/ies and the header file.

This change adds instructions to install the libraries and the header
file and to create cmake files so that dilithium can be easily be used
in other projects.

For using libdilithium in a cmake-based project you can use the following lines:

-  `find_package(dilithium 3.1 REQUIRED)`
- ` target_link_libraries(${TARGET} PRIVATE dilithium2_ref dilithium3_ref
  dilithium5_ref aes256ctr_ref fips202_ref)`
- Additionally the header file needs to be included in the source file `extern
  "C" { #include <dilithium-3.1/api.h> }`