CMakeLists.txt: Enable parallel test execution
The last test depends on the execution of the previos test_vector*
tests as these tests create data, which is hashed and compared to an existing
hash value.

This change adds a dependency for hashes on the test_vector* tests, so that the
test hashes is executed after all of these test_vector tests have been
successfully executed.