Add SPDX headers and copyright notices at the beginning of files.
In downstream projects, files may be copied and moved around. This makes it
difficult to track the copyrights and licenses without having them stored
in each file.

License and copyright holders for files without specific header is assumed
to be the one in LICENSE and AUTHORS.md files respectively.

I haven't touched avx2/f1600x4.S because it is derived from the code at
https://github.com/sphincs/sphincsplus/pull/14/ but there is no copyright
information in that PR and the code is said to be generated by PeachPy but
there is no source code associated.